Refactor: changed type of confidence to double

+++ src/main/java/org/opennars/lab/vision/SymRecognizer.java
@@ -386,7 +386,7 @@ public class SymRecognizer extends javax.swing.JFrame {
                         public void onSolution(Sentence belief) {
                             //System.out.println("solution: " + belief);
                             System.out.println(belief);
-                            float howconf = belief.truth.getConfidence();
+                            double howconf = belief.truth.getConfidence();
                             if(howconf > 0.1f) { //only mark if above 0.1 confidence
                                 //also mark image:
                                 int maxu = Integer.valueOf(belief.getTerm().toString().split("example")[1].split("}")[0]);
